add add to drawer button bar to meta data assets
