+++ src/components/ObjectViewerButtonBar/AddToDrawerButton.vue
