+++ src/components/ObjectViewerButtonBar/AddToEmbeddedPluginButton.vue
